backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
python-jose 
passlib[bcrypt] 
python-multipart
pydantic[email]
pydantic
bcrypt==3.2.2
langgraph==0.1.5
asyncio
aiofiles
requests
supabase
psycopg2-binary
sqlalchemy 
pgvector
streamlit
openai
langfuse    
python-dotenv